+++ CHANGELOG.md
@@ -6,6 +6,12 @@ and this library adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+- The base sapling params download URL has been changed to `https://download.z.cash/downloads/`
+
+### Fixed
+- As part of the sapling params download URL change, the extra `/` character has been removed from the result path
+
 ### Added
 - `SdkSynchronizer.estimateBirthdayHeight(date: Date)` has been added to get an estimated height for a given date,
   typically used for estimating birthday.